Changes for initial tuning of shoulder motor.  Various other cleanup changes at the same time.

+++ src/main/java/org/mayheminc/util/MayhemTalonSRX.java
@@ -135,7 +135,7 @@ public class MayhemTalonSRX extends TalonSRX {
 			this.setSelectedSensorPosition(zeroPositionCount, 0, 1000);
 	}
 
-	public double getPosition() {
+	public int getPosition() {
 		return this.getSelectedSensorPosition(0);
 	}
